LPD-94998 Consolidate validateAPIKey into a shared helper
